Separate out filterCardByMana function
