+++ src/main/Main.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import styled, { css } from 'styled-components';
 import { sortByCMC, sortByColour } from '../utils/sort';
+import { filterCardByMana } from '../utils/filter';
 import { media } from '../utils/theme';
 import Navigation from './navigation/Navigation';
 import Cards from './cards/Cards';
@@ -19,8 +20,6 @@ class Main extends Component {
   }
 
   componentDidMount() {
-    // const baseUrl = 'https://api.scryfall.com/cards/search?q=%28o%3Aflash+or+t%3Ainstant%29+s%3ADOM';
-    // this.fetchCards(baseUrl);
     this.getSets();
     this.storeCards(dummyData);
   }
@@ -72,23 +71,6 @@ class Main extends Component {
     });
   }
 
-  formatManaCostToColorObject = (cost) => {
-    const re = /[\d{}]/g;
-    const formattedCost = cost.replace(re, '');
-    const manaObject = {
-      W: 0,
-      U: 0,
-      B: 0,
-      R: 0,
-      G: 0,
-    };
-    for (let i = 0, n = formattedCost.length; i < n; i += 1) {
-      // add one of the appropriate color
-      manaObject[formattedCost[i]] += 1;
-    }
-    return manaObject;
-  }
-
   filterCards = () => {
     const { cards } = this.state;
     const { mana, sortBy } = this.state.filters;
@@ -96,7 +78,7 @@ class Main extends Component {
     let cardsToShow = cards.slice();
     if (mana) {
       cardsToShow = cardsToShow.filter(card => (
-        this.filterCardByMana(card, mana)
+        filterCardByMana(card, mana)
       ));
     }
     if (sortBy.indexOf('colour') !== -1) {
@@ -108,46 +90,12 @@ class Main extends Component {
     this.setState({ cardsToShow });
   }
 
-  filterCardByMana = (card, mana) => {
-    /**
-     * @param card: the card to be filtered
-     * @param mana: state.filters.mana
-     */
-    // if the mana has a generic number
-    const genericNum = mana.match(/\d/);
-    let filterCMC;
-    if (genericNum) {
-      filterCMC = (genericNum.length - 1) + parseInt(genericNum, 10);
-    } else {
-      filterCMC = mana.length;
-    }
-    if (filterCMC < card.cmc) {
-      return false;
-    }
-    const cardCost = this.formatManaCostToColorObject(card.mana_cost);
-    const manaCost = this.formatManaCostToColorObject(mana);
-    const colors = Object.keys(cardCost);
-    for (let i = 0, n = colors.length; i < n; i += 1) {
-      if (cardCost[colors[i]] > manaCost[colors[i]]) {
-        return false;
-      }
-    }
-    return true;
-  }
-
   toggleGrid = () => {
     this.setState({ isGrid: !this.state.isGrid });
   }
 
-  // fetchCards(url) {
-  //   fetch(url)
-  //     .then(response => response.json())
-  //     .then(data => this.storeCards(data))
-  //     .catch(error => console.log(error));
-  // }
-
   handleSetChange = (val) => {
-    let url = 'https://api.scryfall.com/cards/search?q=%28o%3Aflash+or+t%3Ainstant%29+s%3A'
+    let url = 'https://api.scryfall.com/cards/search?q=%28o%3Aflash+or+t%3Ainstant%29+s%3A';
     url += val;
     this.loadSet(url);
   }
